Convert string based template file to file based
