+++ templates/extension/12/Classes/Constants/GeneralConstants.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace TYPOCONSULT\ExtensionNamePascal\Constants;
+
+class GeneralConstants
+{
+    public const EXT_KEY = 'extensionNameSnake';
+}
