+++ templates/extension/12/Classes/Domain/Model/Content.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace TYPOCONSULT\ExtensionNamePascal\Domain\Model;
+
+class Content extends \TYPOCONSULT\TcSys\Domain\Model\Content
+{
+}
