+++ templates/extension/12/Configuration/Icons.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+use TYPOCONSULT\TcFrontend\Constants\GeneralConstants;
+use TYPOCONSULT\TcTools\Services\IconLoaderService;
+
+$iconLoaderService = new IconLoaderService(GeneralConstants::EXT_KEY);
+
+return $iconLoaderService->readIconsFromFolder()->getIconsForIconsFile();
