+++ templates/extension/12/Configuration/Services.yaml
@@ -1,0 +1,8 @@
+services:
+  _defaults:
+    autowire: true
+    autoconfigure: true
+    public: false
+
+  TYPOCONSULT\ExtensionNamePascal\:
+    resource: '../Classes/*'
