+++ templates/extension/12/Resources/Private/JavaScripts/general.js
@@ -1,0 +1,5 @@
+(function () {
+    document.addEventListener('DOMContentLoaded', function() {
+        console.log("Document ready");
+    });
+}());
