+++ templates/extension/12/Resources/Private/Language/da.locallang_db.xlf
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<xliff version="1.2" xmlns="urn:oasis:names:tc:xliff:document:1.2">
+    <file datatype="plaintext">
+        <header/>
+        <body>
+            <trans-unit id="tt_content.extensionNameClean_plugin.name">
+                <target>Titel</target>
+            </trans-unit>
+            <trans-unit id="tt_content.extensionNameClean_plugin.description">
+                <target>Beskrivelse...</target>
+            </trans-unit>
+        </body>
+    </file>
+</xliff>
