+++ templates/extension/12/Resources/Private/Language/locallang.xlf
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<xliff version="1.2" xmlns="urn:oasis:names:tc:xliff:document:1.2">
+    <file datatype="plaintext">
+        <header/>
+        <body>
+            <trans-unit id="hello.world">
+                <source>Hello World</source>
+            </trans-unit>
+        </body>
+    </file>
+</xliff>
