+++ templates/extension/12/Resources/Private/Language/locallang_db.xlf
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<xliff version="1.2" xmlns="urn:oasis:names:tc:xliff:document:1.2">
+    <file datatype="plaintext">
+        <header/>
+        <body>
+            <trans-unit id="tt_content.extensionNameClean_plugin.name">
+                <source>Title</source>
+            </trans-unit>
+            <trans-unit id="tt_content.extensionNameClean_plugin.description">
+                <source>Description...</source>
+            </trans-unit>
+        </body>
+    </file>
+</xliff>
