+++ templates/extension/12/Resources/Private/Partials/Bodytext.html
@@ -1,0 +1,5 @@
+<f:variable name="bodytext" value="{content.bodytext}"/>
+
+<f:if condition="{bodytext}">
+    {bodytext -> f:format.html()}
+</f:if>
