+++ templates/extension/12/ext_emconf.php
@@ -1,0 +1,6 @@
+<?php
+
+$EM_CONF[$_EXTKEY] = [
+    'title' => 'extensionNamePretty',
+    'state' => 'excludeFromUpdates'
+];
