+++ templates/extension/12/ext_typoscript_constants.typoscript
@@ -1,0 +1,9 @@
+plugin {
+    extensionNamePrefixed {
+        view {
+            templateRootPath = EXT:extensionNameSnake/Resources/Private/Templates/
+            partialRootPath = EXT:extensionNameSnake/Resources/Private/Partials/
+            layoutRootPath = EXT:extensionNameSnake/Resources/Private/Layouts/
+        }
+    }
+}
